Support validUntil in ethers signer
ethers.js rejects any extraneous parameter in a tx, so trying to set a `validUntil` in a tx would result in:

```
Error: cannot override "validUntil" (operation="overrides", overrides=["validUntil"], code=UNSUPPORTED_OPERATION, version=contracts/5.0.2)
```

We had already hit this issue with `speed`, so we made it a signer-level parameter. This change does the same with `validUntil`, but using a `validFor` instead, that calculates the actual `validUntil` when populating the tx (so the same signer can be used over time).

And since we're at it, this also allows setting a default fixed `gasPrice` to the signer.