Add kvstore-client package
Adds another client package for interacting with the kvstore from an autotask. Note that this package, unlike others, can only be used from within an autotask, since we are not exposing an external API to interact with the kvstore (for now).

Also extracts lambda invoking out of the relayer package and into the base package, for reusability.

Pending:
- Write README
- Type and max size validation for keys and values